requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
httpx>=0.27.0
mlx>=0.17.0
mlx-lm>=0.18.0
faiss-cpu>=1.8.0
numpy>=1.26.0
huggingface_hub>=0.24.0
ollama>=0.4.0
# sqlite3 is included in Python standard library
